add delete only users own bookmarks

+++ lib/bookmark_web/controllers/page_controller.ex
@@ -12,18 +12,29 @@ defmodule BookmarkWeb.PageController do
   end
 
   def delete(conn, params) do
-    query = Repo.get_by(Shortcode, id: params["to_be_deleted"])
-    Repo.delete(query)
+    user_id = conn.assigns.current_user.id
+    ptbd = Repo.get(Shortcode, params["to_be_deleted"])
+
+    if ptbd.created_by == user_id do
+      query = Repo.get_by(Shortcode, id: params["to_be_deleted"])
+      Repo.delete(query)
+    else
+      conn
+      |> put_flash(:error, "Don't touch that it's not yours to delete.")
+      |> redirect(to: "/")
+    end
 
     redirect(conn, to: Routes.page_path(conn, :index))
-
   end
 
   def create(conn, params) do
+    IO.inspect(conn.assigns.current_user.id, label: "user id")
+
     code = Helpers.generate()
+    user_id = conn.assigns.current_user.id
 
     %Shortcode{}
-    |> Shortcode.changeset(%{"url" => params["url"], "code" => code})
+    |> Shortcode.changeset(%{"url" => params["url"], "code" => code, "created_by" => user_id})
     |> Repo.insert()
 
     redirect(conn, to: Routes.page_path(conn, :index))
